Function that prints fibonacci numbers sum of even

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,17 +9,17 @@
 int main(void)
 {
 	int i;
-	long int t1, t2, nextTerm;
+	unsigned long long t1, t2, nextTerm;
 
 	t1 = 1;
 	t2 = 2;
 	nextTerm = t1 + t2;
 
-	printf("%li, %li", t1, t2);
+	printf("%lu, %lu", t1, t2);
 
 	for (i = 3; i <= 98; ++i)
 	{
-		printf(", %li", nextTerm);
+		printf(", %lu", nextTerm);
 		t1 = t2;
 		t2 = nextTerm;
 		nextTerm = t1 + t2;
